Fixed the TODO again, forgot to change datatype

+++ src/DataTypes.java
@@ -5,7 +5,7 @@ public class DataTypes {
 
         long s = 0;
         // below is a "foreach" loop which iterates through numbers
-        for (int x : numbers) {
+        for (long x : numbers) {
             s += x;
         }
         return s;
